Moves auth tests to its own file
